Moved old createCollection code from ClusterInfo to testOnly land. Unfortunately some tests use it, and we cannot trivially make them use official apis
